Scheduler/requirements.txt
APScheduler==3.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
google-auth==2.3.2
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.20.1
idna==3.3
oauth2client==4.1.3
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
twilio==7.2.0
tzdata==2021.5
tzlocal==4.0.2
urllib3==1.26.7
